Extract shared checkpoint mixin; drop hard-coded test data path
Address PR #76 quality review: TransitionModel duplicated
PolicyValueNetwork's save/load pair, so both now use a shared
TorchCheckpointMixin whose classmethod load keeps cls bound to the
concrete subclass (public APIs unchanged). The partial-data smoke test
hard-coded an absolute path to one machine; it now resolves the
directory repo-relative and honors a BICAMERAL_BASELINE_DATA override,
keeping the skipif guard for machines without the pilot episodes.

Refs #27.
